Migrate syntax of JSON graphs
